tests/wpt/tests/html/canvas/offscreen/transformations/2d.transformation.translate.basic.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.transformation.translate.basic</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.transformation.translate.basic</h1>
<p class="desc">translate() works</p>


<script>
var t = async_test("translate() works");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  ctx.fillStyle = '#f00';
  ctx.fillRect(0, 0, 100, 50);

  ctx.translate(100, 50);
  ctx.fillStyle = '#0f0';
  ctx.fillRect(-100, -50, 100, 50);
  _assertPixel(canvas, 90,40, 0,255,0,255);
  t.done();

});
</script>
